[#32] refactor: 리뷰 반영 리팩토링

+++ src/main/java/com/attica/athens/domain/agora/dto/SimpleParticipants.java
@@ -1,8 +1,8 @@
 package com.attica.athens.domain.agora.dto;
 
 public record SimpleParticipants(
-    long pros,
-    long cons,
-    long observer
+    int pros,
+    int cons,
+    int observer
 ) {
 }
